requirements.txt
bidict==0.22.1
connected_components_3d==3.18.0
easydict==1.10
einops==0.8.0
hydra-core==1.3.2
matplotlib==3.7.3
monai==1.3.0
nibabel==5.3.2
numpy==2.1.3
omegaconf==2.3.0
pandas==1.5.3
pytorch_lightning==1.8.6
PyYAML==6.0.2
scikit_learn==1.2.2
scikit_survival==0.21.0
torch==2.4.0 # download for cuda via pip from pytorch.org
torchmetrics==1.3.2
wandb==0.16.5
